change reset-password/send-code/ method

+++ apps/authentication/views.py
@@ -95,6 +95,10 @@ class ResetSendCodeAPIView(generics.CreateAPIView):
         serializer.is_valid(raise_exception=True)
         email = request.data['email']
 
+        if User.objects.filter(email='contact-center.burarrc@mail.ru'):
+            return Response({"result": False, "email": ['Не удалось отправить код']},
+                            status=status.HTTP_404_NOT_FOUND)
+
         if User.objects.filter(email=email).exists():
             if ResetEmailCode.objects.filter(email=email).exists():
                 ResetEmailCode.objects.filter(email=email).delete()
